renamed beta to project
