+++ cmd/flow/main.go
@@ -24,13 +24,13 @@ import (
 
 	cli "github.com/onflow/flow-cli/flow"
 	"github.com/onflow/flow-cli/flow/accounts"
-	"github.com/onflow/flow-cli/flow/beta"
 	"github.com/onflow/flow-cli/flow/blocks"
 	"github.com/onflow/flow-cli/flow/cadence"
 	"github.com/onflow/flow-cli/flow/collections"
 	"github.com/onflow/flow-cli/flow/emulator"
 	"github.com/onflow/flow-cli/flow/initialize"
 	"github.com/onflow/flow-cli/flow/keys"
+	"github.com/onflow/flow-cli/flow/project"
 	"github.com/onflow/flow-cli/flow/scripts"
 	"github.com/onflow/flow-cli/flow/transactions"
 	"github.com/onflow/flow-cli/flow/version"
@@ -42,7 +42,7 @@ var cmd = &cobra.Command{
 }
 
 func init() {
-	cmd.AddCommand(beta.Cmd)
+	cmd.AddCommand(project.Cmd)
 	cmd.AddCommand(initialize.Cmd)
 	cmd.AddCommand(accounts.Cmd)
 	cmd.AddCommand(blocks.Cmd)
